third_party/WebKit/LayoutTests/html/tabular_data/td_rowspan.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<script>
description("HTMLTableCellElement rowspan attribute test");

function rowspanAttributeEffect(value)
{
    var element = document.createElement("td");
    element.setAttribute("rowSpan", value);
    return element.rowSpan;
}

shouldBe('document.createElement("td").rowSpan', '1');

shouldBe('rowspanAttributeEffect("")', '1');

shouldBe('rowspanAttributeEffect("1")', '1');
shouldBe('rowspanAttributeEffect("2")', '2');
shouldBe('rowspanAttributeEffect("10")', '10');

shouldBe('rowspanAttributeEffect("0")', '0');

shouldBe('rowspanAttributeEffect("-1")', '1');

shouldBe('rowspanAttributeEffect("1x")', '1');
shouldBe('rowspanAttributeEffect("1.")', '1');
shouldBe('rowspanAttributeEffect("1.9")', '1');
shouldBe('rowspanAttributeEffect("2x")', '2');
shouldBe('rowspanAttributeEffect("2.")', '2');
shouldBe('rowspanAttributeEffect("2.9")', '2');

shouldBe('rowspanAttributeEffect("a")', '1');
shouldBe('rowspanAttributeEffect("\v7")', '1');
shouldBe('rowspanAttributeEffect("  7")', '7');

var arabicIndicDigitOne = String.fromCharCode(0x661);
shouldBe('rowspanAttributeEffect(arabicIndicDigitOne)', '1');
shouldBe('rowspanAttributeEffect("2" + arabicIndicDigitOne)', '2');

shouldBe('rowspanAttributeEffect("2147483647")', '65534');
shouldBe('rowspanAttributeEffect("4294967295")', '65534');
shouldBe('rowspanAttributeEffect("4294967296")', '65534');
</script>
</body>
</html>
