third_party/blink/web_tests/platform/generic/html/tabular_data/td_rowspan-expected.txt

HTMLTableCellElement rowspan attribute test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.createElement("td").rowSpan is 1
PASS rowspanAttributeEffect("") is 1
PASS rowspanAttributeEffect("1") is 1
PASS rowspanAttributeEffect("2") is 2
PASS rowspanAttributeEffect("10") is 10
PASS rowspanAttributeEffect("0") is 0
PASS rowspanAttributeEffect("-1") is 1
PASS rowspanAttributeEffect("1x") is 1
PASS rowspanAttributeEffect("1.") is 1
PASS rowspanAttributeEffect("1.9") is 1
PASS rowspanAttributeEffect("2x") is 2
PASS rowspanAttributeEffect("2.") is 2
PASS rowspanAttributeEffect("2.9") is 2
PASS rowspanAttributeEffect("a") is 1
PASS rowspanAttributeEffect("7") is 1
PASS rowspanAttributeEffect("  7") is 7
PASS rowspanAttributeEffect(arabicIndicDigitOne) is 1
PASS rowspanAttributeEffect("2" + arabicIndicDigitOne) is 2
PASS rowspanAttributeEffect("2147483647") is 65534
PASS rowspanAttributeEffect("4294967295") is 65534
PASS rowspanAttributeEffect("4294967296") is 65534
PASS successfullyParsed is true

TEST COMPLETE
